third_party/blink/web_tests/external/wpt/service-workers/service-worker/interface-requirements-sw.https-expected.txt

This is a testharness.js-based test.
[PASS] Interfaces and attributes in ServiceWorkerGlobalScope
[FAIL] Event constructors
  assert_equals: FetchEvent.isReload should not exist expected (undefined) undefined but got (boolean) false
[PASS] xhr is not exposed
[PASS] URL.createObjectURL is not exposed
Harness: the test ran to completion.
